Add 1

+++ app/Http/Controllers/PaymentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Session;
 use DB;
+use Mail;
 use Cart;
 use App\Order;
 use App\Payment;
@@ -13,36 +14,36 @@ use Carbon\Carbon;
 
 class PaymentController extends Controller
 {
-  public function payment_page()
-  {
-    return view('frontEnd.payment.paymentContent');
-  }
-  public function payment_method_save(Request $req)
-  {
-    $payment_type = $req->payment_type;
-    $customer_id = Session::get('customer_id');
-    $shipping_id = Session::get('shipping_id');
-    $order_total = Cart::total();
-    $order_total= str_replace(',','',$order_total);
+    public function payment_page()
+    {
+        return view('frontEnd.payment.paymentContent');
+    }
+    public function payment_method_save(Request $req)
+    {
+        $payment_type = $req->payment_type;
+        $customer_id = Session::get('customer_id');
+        $shipping_id = Session::get('shipping_id');
+        $order_total = Cart::total();
+        $order_total= str_replace(',', '', $order_total);
 
-    $order_id = Order::insertGetId([
+        $order_id = Order::insertGetId([
       'customer_id'=>$customer_id,
       'shipping_id'=>$shipping_id,
       'order_total'=>$order_total,
       'created_at'=>Carbon::now(),
     ]);
 
-    $payment_id =Payment::insertGetId([
+        $payment_id =Payment::insertGetId([
       'order_id'=>$order_id,
       'payment_type'=>$payment_type,
       'created_at'=>Carbon::now(),
     ]);
 
-    $cart = Cart::content();
+        $cart = Cart::content();
 
 
-    foreach ($cart as $cart) {
-    Orderdetail::insert([
+        foreach ($cart as $cart) {
+            Orderdetail::insert([
     'order_id'=>$order_id,
     'product_id'=> $cart->id,
     'product_name'=>$cart->name,
@@ -50,28 +51,42 @@ class PaymentController extends Controller
     'product_sales_quantity'=>$cart->qty,
     'created_at'=>Carbon::now(),
       ]);
+        }
+
+        if (!empty($order_total)) {
+            if ($payment_type == 'cash') {
+                $data = array('name'=>'Jabir');
+                // Mail::send('mail', $data, function ($message) {
+                //     $message->to('abc@gmail.com', 'Tutorials Point')->subject('Laravel Testing Mail with Attachment');
+                //     $message->attach('C:\laravel-master\laravel\public\uploads\image.png');
+                //     $message->attach('C:\laravel-master\laravel\public\uploads\test.txt');
+                //     $message->from('xyz@gmail.com', 'Virat Gandhi');
+                // });
+                Mail::send('mail.mail', $data, function ($message) {
+                    $message->to('jabirhasan88@gmail.com', 'My Point')->subject('Laravel HTML Testing Mail');
+                    $message->from('apon@gmail.com', 'Apon');
+                });
+
+
+                Cart::destroy();
+                return redirect()->back()->with('success', 'Payment success');
+            } elseif ($payment_type == 'bkash') {
+                return redirect('/addmoney/stripe');
+            // return "Bkash is Not available";
+    // code...
+            } else {
+                return redirect('/paypal');
+            }
+        } else {
+            echo "Amr Matha";
+        }
     }
 
-  if ($payment_type == 'cash') {
-    Cart::destroy();
-    return redirect()->back()->with('success','Payment success');
-  }elseif ($payment_type == 'bkash') {
-    return redirect('/addmoney/stripe');
-    // return "Bkash is Not available";
-    // code...
-  }else {
-    return redirect('/paypal');
-  }
 
-
-  }
-
-
-  public function invoice($order_id)
-  {
-
-    $Customer_details = DB::table('orders')
-          ->where('order_id',$order_id)
+    public function invoice($order_id)
+    {
+        $Customer_details = DB::table('orders')
+          ->where('order_id', $order_id)
           ->join('customers', 'orders.customer_id', '=', 'customers.customer_id')
           ->join('shippings', 'orders.shipping_id', '=', 'shippings.shipping_id')
 
@@ -81,17 +96,17 @@ class PaymentController extends Controller
 
 
 
-    $order_details =DB::table('orderdetails')
+        $order_details =DB::table('orderdetails')
 
           ->join('orders', 'orderdetails.order_id', '=', 'orders.order_id')
           ->select('orderdetails.*', 'orders.*')
-          ->get()->where('order_id',$order_id);
+          ->get()->where('order_id', $order_id);
 
-          // echo "<pre>";
-          // print_r($order_details);
-          // echo "</pre>";
-          // return view('admin.order.orderview',);
+        // echo "<pre>";
+        // print_r($order_details);
+        // echo "</pre>";
+        // return view('admin.order.orderview',);
 
-    return view('admin.invoice.invoiceContent',compact('order_details','Customer_details'));
-  }
+        return view('admin.invoice.invoiceContent', compact('order_details', 'Customer_details'));
+    }
 }
